Create team_code_share_seasonal_av.py
I release particles globally on a 2x2 degree grid at the start of each season (DJF, MAM, JJA, SON) and run it for 90 days. I repeat this for 10 years (i.e. I have 40 runs).  I see the effects of biofouling on the sinking of plastics  so here I plot the 10-year average settling onset time (when the particles first start to sink) for each season. I am still in the learning phase for python so thanks for any help!